Major refactoring of Shupito connections in preparation for USB support.
There are bunch of outputs that each refcount debugging. Also,
there is a major architectural flaw in how connection removal is handled.
I failed to realize that when QObject::destroyed signal is emitted,
the object's dynamic type is QObject and any super-objects are already
destroyed. As such there are some invalid static_casts (they will not
cause a crash, but should be removed anyway).